+++ syslog2irc/irc.py
@@ -31,7 +31,8 @@ class Channel(namedtuple('Channel', 'name password')):
 class Bot(SingleServerIRCBot):
     """An IRC bot to forward syslog messages to IRC channels."""
 
-    def __init__(self, server_spec, nickname, realname, channels, ssl=False):
+    def __init__(self, server_spec, nickname, realname, channels, ssl=False,
+                 shutdown_predicate=None):
         log('Connecting to IRC server {0.host}:{0.port:d} ...', server_spec)
 
         connect_params = {}
@@ -44,6 +45,8 @@ class Bot(SingleServerIRCBot):
 
         # Note: `self.channels` already exists in super class.
         self.channels_to_join = channels
+
+        self.shutdown_predicate = shutdown_predicate
 
     def get_version(self):
         """Return this on CTCP VERSION requests."""
@@ -79,14 +82,12 @@ class Bot(SingleServerIRCBot):
         log('Cannot join channel {} (bad key).', channel_name)
 
     def on_privmsg(self, conn, event):
-        """React on private messages.
-
-        Shut down, for example.
-        """
+        """React on private messages."""
         nickmask = event.source
         text = event.arguments[0]
 
-        if message == 'shutdown!':
+        if self.shutdown_predicate is not None \
+                and self.shutdown_predicate(nickmask, text):
             self.shutdown(nickmask)
 
     def shutdown(self, nickmask):